requirements.txt
pymongo[srv]
flask
flask-wtf
flask-sqlalchemy
flask-migrate
Werkzeug
flask-login